Revise the positional embedding API

+++ tests/common.py
@@ -21,7 +21,7 @@ device = Device("cpu")
 def assert_close(a: Tensor, b: Union[Tensor, List[Any]]) -> None:
     """Asserts that ``a`` and ``b`` are element-wise equal within a tolerance."""
     if not isinstance(b, Tensor):
-        b = torch.tensor(b, device=device)
+        b = torch.tensor(b, device=device, dtype=a.dtype)
 
     torch.testing.assert_close(a, b)  # type: ignore[attr-defined]
 
@@ -29,7 +29,7 @@ def assert_close(a: Tensor, b: Union[Tensor, List[Any]]) -> None:
 def assert_equal(a: Tensor, b: Union[Tensor, List[Any]]) -> None:
     """Asserts that ``a`` and ``b`` are element-wise equal."""
     if not isinstance(b, Tensor):
-        b = torch.tensor(b, device=device)
+        b = torch.tensor(b, device=device, dtype=a.dtype)
 
     torch.testing.assert_close(a, b, rtol=0, atol=0)  # type: ignore[attr-defined]
